Fix some KeePassXC-Browser support
This allows KeePassXC to configure browser extensions, providing that
a supported browser is either installed without sandboxing or shares
the required configuration directories outside of its sandbox.

In addition to this, KeePassXC is patched to use the correct runtime
directory location for its UNIX socket (kpxc_server).

Ultimately, with this change and some other manual workarounds it is
possible to get KeePassXC-Browser working.

For more information, see:
https://github.com/flathub/org.keepassxc.KeePassXC/issues/13